Add files via upload
The structure of the code is as follows:
In the file "E.py " the is the definition of the Heisenberg Hamiltonian and the momentum operator. 
The "matrix_plot.py" can by used to create 2d plots from the matrices in "E.py".
The "circuit.py" has definitions of the circuits that can be used for the computations.
The "angles.py" includes the definition of the angle arrays used by the circuit. 
The "states.py" creates the "state_dict.py" file which contains the eigenvalues and eigenvectors of the Heisenberg Hamiltonian.
The "b_values.py" creates the "b_values_data.py" which contains the b arrays used for the computation of excited states.
The "optimizing.py" is the main file for optimization.
The "optimizing_functions.py" contains the definition of functions used for the computation of the ground and excited states.
The "custom_optimizers.py" contains a function for sequencial optimization.
The "main.py" is the central file through which the user may interact. 
The resulting plots and data are created in the /trial folder.

Please note that the code is currently under heavy customization.
The code is made to test arXiv:2002.06210v2 and arXiv:1805.08138v5